ca-geo-weather-daily/requirements.txt
requests>=2.28.0,<3
snowflake-connector-python>=3.0.0,<4
cryptography>=41.0.0
